Added all the files for this project also .zip file
The zip file contains all the files for this project (including the original formatted README.txt document).  This is done because to use this project, the folder within the zip (containing all the add-in files) must be put directly into the Microsoft Add-Ins folder and its name must not be changed!  The individual files found within the only directory of the zip file have also been added to this repository so they can be pulled/read separately